Added TxtHandler, FileParser, FileHandler interface
Math10.txt was used to test the TxtHandler.  The .txt file was manually
pre-formatted so other parts of the project could be implemented that
depend on the class.  I'm currently implementing functions for
non-preformatted text files which will be added to TxtHandler Friday.
Relationships need to be added to the FileParser class.